Splitting item definitins into their own files.
